requirements.txt
requests
python-dotenv
solana
aiohttp
websockets
pandas